RavenDB-14931 fix CanIndexRollups test
