+++ test/SlowTests/Issues/RavenDB_14932.cs
@@ -46,8 +46,12 @@ namespace SlowTests.Issues
                     PolicyCheckFrequency = TimeSpan.FromSeconds(1)
                 };
                 await store.Maintenance.SendAsync(new ConfigureTimeSeriesOperation(config));
-
+                var database = await GetDocumentDatabaseInstanceFor(store);
                 var now = DateTime.UtcNow;
+                var nowSeconds = now.Second;
+                now = now.AddSeconds(-nowSeconds);
+                database.Time.UtcDateTime = () => DateTime.UtcNow.AddSeconds(-nowSeconds);
+
                 var baseline = now.AddDays(-12);
                 var total = ((TimeSpan)TimeValue.FromDays(12)).TotalMinutes;
 
@@ -64,9 +68,11 @@ namespace SlowTests.Issues
                     session.SaveChanges();
                 }
 
-                var database = await GetDocumentDatabaseInstanceFor(store);
-                await database.TimeSeriesPolicyRunner.RunRollups();
-                await database.TimeSeriesPolicyRunner.DoRetention();
+                await WaitForPolicyRunner(database);
+                await VerifyPolicyExecution(store, config.Collections["Users"], 4, policies: new List<TimeSeriesPolicy> { p1 });
+                await VerifyPolicyExecution(store, config.Collections["Users"], 5, policies: new List<TimeSeriesPolicy> { p2 });
+                await VerifyPolicyExecution(store, config.Collections["Users"], 2, policies: new List<TimeSeriesPolicy> { p3 });
+                await VerifyPolicyExecution(store, config.Collections["Users"], 3, policies: new List<TimeSeriesPolicy> { p4 });
 
                 WaitForIndexing(store);
                 RavenTestHelper.AssertNoIndexErrors(store);
@@ -74,6 +80,7 @@ namespace SlowTests.Issues
                 using (var session = store.OpenSession())
                 {
                     var user = session.Load<User>("users/karmel");
+
                     var count = session
                         .TimeSeriesFor(user, "Heartrate")
                         .Get(DateTime.MinValue, DateTime.MaxValue)
@@ -102,7 +109,7 @@ namespace SlowTests.Issues
                     var results = session.Query<TimeSeriesIndex.Result, TimeSeriesIndex>()
                         .ToList();
 
-                    Assert.Equal(count, results.Count);
+                    Assert.True(count == results.Count, $"Test time = {now}");
                 }
             }
         }
